+++ react-ui/src/components/addOrder/AddSample.js
@@ -1,42 +1,50 @@
 import React, {useEffect, useRef, useState} from 'react'
-import {Button, Column, Grid, Row} from '@carbon/react';
-import SampleTypes from './SampleTypes';
+import {Button, Link, Row, Stack} from "@carbon/react";
+import {Add} from '@carbon/react/icons';
 import {getFromOpenElisServer} from "../utils/Utils";
+import SampleType from "./SampleType";
 
 const AddSample = (props) => {
+    const {samples, setSamples} = props;
     const componentMounted = useRef(true);
-    const [sampleElementsList, setSampleElementsList] = useState([{index: "", tests: [],referralItems:[]}]);
     const [elementsCounter, setElementsCounter] = useState(0);
+
     const [rejectSampleReasons, setRejectSampleReasons] = useState([]);
 
     const handleAddNewSample = () => {
-        setElementsCounter(elementsCounter + 1);
-        const updates = [...sampleElementsList];
-        updates.push({
-            index: elementsCounter, tests: [], referralItems: []
-        });
-        setSampleElementsList(updates);
+        let updateSamples = [...samples];
+        let count = elementsCounter + 1
+        updateSamples.push({
+            index: count,
+            referralItems: [],
+            sampleTypeId: "",
+            sampleXML: null,
+            tests: []
+        })
+        setSamples(updateSamples);
+        setElementsCounter(count);
     }
 
-    function sampleTypeObject(object) {
-        const {sampleTypeId, selectedTests, sampleXML, referralItems, index} = object;
-        let newState = [...sampleElementsList];
-        if (selectedTests && selectedTests.length > 0) {
-            newState[index].sampleTypeId = sampleTypeId;
-            newState[index].tests = selectedTests;
+    const sampleTypeObject = (object) => {
+
+        const {sampleTypeId, selectedTests, sampleXML, referralItems, sampleObjectIndex} = object;
+        let newState = [...samples];
+        if (sampleTypeId) {
+            console.log(samples)
+            newState[sampleObjectIndex].sampleTypeId = sampleTypeId;
+        } else if (selectedTests && selectedTests.length > 0) {
+            newState[sampleObjectIndex].tests = selectedTests;
         } else if (referralItems && referralItems.length > 0) {
-            newState[index].referralItems = referralItems;
+            newState[sampleObjectIndex].referralItems = referralItems;
         } else if (sampleXML != null) {
-            newState[index].sampleXML = sampleXML;
+            newState[sampleObjectIndex].sampleXML = sampleXML;
         }
-        setSampleElementsList(newState);
-        props.setSamples(sampleElementsList);
+        props.setSamples(newState);
     }
 
     const removeSample = (index) => {
-        let newList = sampleElementsList.splice(index, 1);
-        setSampleElementsList(newList);
-        props.setSamples(sampleElementsList);
+        let updateSamples = samples.splice(index, 1);
+        setSamples(updateSamples);
     }
 
     const fetchRejectSampleReasons = (res) => {
@@ -45,42 +53,52 @@ const AddSample = (props) => {
         }
     }
 
-    function initialiseSampleElements() {
-        let newState = [...sampleElementsList];
-        newState[0].index = elementsCounter;
-        newState[0].sampleTypeId = "";
-        newState[0].sampleXML = null;
-        newState[0].tests = [];
-        newState[0].referralItems = [];
-        setSampleElementsList(newState);
-        setElementsCounter(elementsCounter + 1);
+
+    const handleRemoveSample = (e, sample) => {
+        e.preventDefault();
+        let filtered = samples.filter(function (element) {
+            return element !== sample
+        });
+        setSamples(filtered);
     }
 
-
     useEffect(() => {
-        initialiseSampleElements();
         getFromOpenElisServer("/rest/test-rejection-reasons", fetchRejectSampleReasons);
         return () => {
             componentMounted.current = false
         }
     }, []);
 
-    return (<>
-        <Grid fullWidth={true}>
-            <Column lg={16}>
-                {sampleElementsList && sampleElementsList.map((element, index) => {
-                    return (<SampleTypes index={index} key={index} rejectSampleReasons={rejectSampleReasons}
-                                         removeSample={removeSample} sampleTypeObject={sampleTypeObject}/>)
-                })}
-                <Row>
-                    <div className="inlineDiv">
-                        <Button onClick={handleAddNewSample}>Add New Sample</Button>
-                    </div>
-                </Row>
-            </Column>
-        </Grid>
 
-    </>)
+    return (
+        <>
+            <h3>SAMPLE</h3>
+            <Stack gap={10}>
+                <div className="orderLegendBody">
+                    {samples.map((sample, i) => {
+                        return (
+                            <div className="sampleType" key={i}>
+                                <h4>Sample {i + 1}</h4>
+                                <Link href="#" onClick={(e) => handleRemoveSample(e, sample)}>Remove
+                                    Sample</Link>
+                                <SampleType index={i} rejectSampleReasons={rejectSampleReasons}
+                                            removeSample={removeSample} sample={sample}
+                                            sampleTypeObject={sampleTypeObject}/>
+                            </div>
+                        )
+                    })}
+                    <Row>
+                        <div className="inlineDiv">
+                            <Button onClick={handleAddNewSample}>Add Sample
+                                &nbsp;  &nbsp;
+                                <Add size={16}/>
+                            </Button>
+                        </div>
+                    </Row>
+                </div>
+            </Stack>
+        </>
+    )
 }
 
 export default AddSample;
