--- react-ui/src/components/addOrder/SamplePayLoad.js
@@ -1,143 +1,0 @@
-export const samplePayLoad = {
-    rememberSiteAndRequester: false,
-    currentDate: null,
-    projects: null,
-    customNotificationLogic: true,
-    patientEmailNotificationTestIds: [],
-    patientSMSNotificationTestIds: [],
-    providerEmailNotificationTestIds: [],
-    providerSMSNotificationTestIds: [],
-    patientUpdateStatus: "ADD",
-    referralItems: [
-        {
-            referralId: null,
-            accessionNumber: null,
-            sampleType: null,
-            referringTestName: "",
-            referralResults: "",
-            referralStatus: null,
-            referralDate: null,
-            referrer: "",
-            referredInstituteId: "",
-            referredTestId: "",
-            referredTestIdShadow: null,
-            testSelectionList: null,
-            referredResult: "",
-            referredDictionaryResult: null,
-            referredMultiDictionaryResult: "",
-            referredResultType: "",
-            dictionaryResults: [],
-            referredSendDate: "",
-            referredReportDate: null,
-            pastNotes: null,
-            note: null,
-            additionalTestsXMLWad: null,
-            referralReasonId: "",
-            referralResultId: null,
-            modified: false,
-            additionalTests: null,
-            inLabResultId: null,
-            multiSelectResultValues: null,
-            qualifiedResultValue: null
-        }
-    ],
-    referralOrganizations: null,
-    referralReasons: null,
-    sampleTypes: null,
-    sampleXML: "",
-    patientProperties: {
-        currentDate: "",
-        patientLastUpdated: "",
-        personLastUpdated: "",
-        patientUpdateStatus: "UPDATE",
-        patientPK: "",
-        STnumber: null,
-        subjectNumber: "",
-        nationalId: "",
-        guid: "",
-        lastName: "",
-        firstName: "",
-        aka: null,
-        mothersName: null,
-        mothersInitial: null,
-        streetAddress: "",
-        city: "",
-        commune: "",
-        addressDepartment: null,
-        gender: "",
-        ageYears: null,
-        ageMonths: null,
-        ageDays: null,
-        birthDateForDisplay: "",
-        patientType: "",
-        insuranceNumber: null,
-        occupation: null,
-        primaryPhone: "",
-        email: null,
-        healthRegion: "",
-        education: "",
-        maritialStatus: "",
-        nationality: "",
-        healthDistrict: null,
-        otherNationality: "",
-        patientContact: null,
-        readOnly: false,
-        patientIdentities: null
-    },
-    patientSearch: null,
-    patientEnhancedSearch: null,
-    patientClinicalProperties: null,
-    sampleOrderItems: {
-        newRequesterName: "",
-        orderTypes: null,
-        orderType: null,
-        externalOrderNumber: null,
-        labNo: "",
-        requestDate: "",
-        receivedDateForDisplay: "",
-        receivedTime: "",
-        nextVisitDate: "",
-        requesterSampleID: null,
-        referringPatientNumber: null,
-        referringSiteId: null,
-        referringSiteDepartmentId: null,
-        referringSiteCode: null,
-        referringSiteName: "",
-        referringSiteDepartmentName: null,
-        referringSiteList: null,
-        referringSiteDepartmentList: null,
-        providersList: null,
-        providerId: null,
-        providerPersonId: null,
-        providerFirstName: "",
-        providerLastName: "",
-        providerWorkPhone: "",
-        providerFax: "",
-        providerEmail: "",
-        facilityAddressStreet: null,
-        facilityAddressCommune: null,
-        facilityPhone: null,
-        facilityFax: null,
-        paymentOptionSelection: "",
-        paymentOptions: null,
-        modified: true,
-        sampleId: "",
-        readOnly: false,
-        billingReferenceNumber: "",
-        testLocationCode: "",
-        otherLocationCode: "",
-        testLocationCodeList: null,
-        program: "",
-        programList: null,
-        contactTracingIndexName: null,
-        contactTracingIndexRecordNumber: null,
-        priorityList: null,
-        priority: ""
-    },
-    initialSampleConditionList: [],
-    sampleNatureList: null,
-    testSectionList: null,
-    warning: false,
-    useReferral: true,
-    rejectReasonList: null
-}
