+++ react-ui/src/components/layout/Header.js
@@ -218,7 +218,7 @@ class OEHeader extends React.Component {
                           <SideNavMenuItem href="/admin">Admin</SideNavMenuItem>
 
                           <SideNavMenuItem target="_blank" href={"http://ozone.uwdigi.org:8069/"}><FormattedMessage id="admin.billing" /></SideNavMenuItem>
-
+                          <SideNavMenuItem href="/testForm">testFormss</SideNavMenuItem>
                         </SideNavItems>
                       </SideNav>
                     </>
